Consistent naming of Quat and Vector classes
