+++ src/main/java/org/terasology/math/geom/Quat4d.java
@@ -21,147 +21,197 @@ package org.terasology.math.geom;
  * point x,y,z,w coordinates.
  * @author Martin Steiger
  */
-public abstract class Quat4d {
-    
-    /**
-     * The immutable identity quaternion
-     */
-    public static final ImmutableQuat4d IDENTITY = new ImmutableQuat4d(0, 0, 0, 1);
+public class Quat4d extends BaseQuat4d {
 
-    private static final double EPS = 1.0e-12;
-    
-    /**
-     * @return the x component
-     */
-    public abstract double getX();
-    
-    /**
-     * @return the y component
-     */
-    public abstract double getY();
-    
-    /**
-     * @return the z component
-     */
-    public abstract double getZ();
-    
-    /**
-     * @return the w component
-     */
-    public abstract double getW();
+    private double x;
+    private double y;
+    private double z;
+    private double w;
 
     /**
-     * @return the rotation angle
+     * Constructs and initializes a Quat4d from the specified Quat4d.
+     * @param other the Quat4d containing the initialization x y z w data
      */
-    public double getAngle() {
-        return 2 * Math.acos(getW());
+    public Quat4d(BaseQuat4d other) {
+        this.x = other.getX();
+        this.y = other.getY();
+        this.z = other.getZ();
+        this.w = other.getW();
     }
-    
+
     /**
-     * @return a copy of the axis components
+     * Constructs and initializes a Quat4d from the specified xyzw coordinates.
+     * @param x the x coordinate
+     * @param y the y coordinate
+     * @param z the z coordinate
+     * @param w the w scalar component
      */
-    public Vector3d getAxis() {
-        return new Vector3d(getX(), getY(), getZ());
+    public Quat4d(double x, double y, double z, double w) {
+        double mag = 1.0 / Math.sqrt(x * x + y * y + z * z + w * w);
+        this.x = x * mag;
+        this.y = y * mag;
+        this.z = z * mag;
+        this.w = w * mag;
     }
-    
+
     /**
-     * Performs a great circle interpolation between quaternion q1
-     * and quaternion q2 and places the result into this quaternion.
-     * @param q1  the first quaternion
-     * @param q2  the second quaternion
-     * @param alpha  the alpha interpolation parameter
-     * @return the interpolated quaternion
+     * Constructs and initializes a Quat4d from the array of length 4. 
+     * @param q the array of length 4 containing xyzw in order
      */
-    public static final MutableQuat4d interpolate(Quat4d q1, Quat4d q2, double alpha) {
-        // From "Advanced Animation and Rendering Techniques"
-        // by Watt and Watt pg. 364, function as implemented appeared to be
-        // incorrect.  Fails to choose the same quaternion for the double
-        // covering. Resulting in change of direction for rotations.
+    public Quat4d(double[] q) {
+        this(q[0], q[1], q[2], q[3]);
+    }
 
-        double dot = q2.getX() * q1.getX() + q2.getY() * q1.getY() + q2.getZ() * q1.getZ() + q2.getW() * q1.getW();
+    /** 
+     * Constructs and initializes a Quat4d from the specified Tuple4d.  
+     * @param t the Tuple4d containing the initialization x y z w data 
+     */
+    public Quat4d(BaseVector4d t) {
+        this(t.getX(), t.getY(), t.getZ(), t.getW());
+    }
 
-        double q1x;
-        double q1Y;
-        double q1Z;
-        double q1W;
-        
-        // Negate the first quaternion in the case that the
-        // dot product of q1 and this is negative.
-        if (dot < 0) {
-            q1x = -q1.getX();
-            q1Y = -q1.getY();
-            q1Z = -q1.getZ();
-            q1W = -q1.getW();
-            dot = -dot;
+    /**
+     * @param axis the axis. Length must be != 0
+     * @param angle the rotation angle in radians
+     */
+    public Quat4d(Vector3d axis, double angle) {
+        double d = axis.length();
+        double s = Math.sin(angle * 0.5) / d;
+        x = axis.getX() * s;
+        y = axis.getY() * s;
+        z = axis.getZ() * s;
+        w = Math.cos(angle * 0.5);
+    }
+
+    /**
+     * @param yaw the yaw angle (in radians)
+     * @param pitch the pitch angle (in radians)
+     * @param roll the roll angle (in radians)
+     */
+    public Quat4d(double yaw, double pitch, double roll) {
+        double halfYaw = yaw * 0.5f;
+        double halfPitch = pitch * 0.5f;
+        double halfRoll = roll * 0.5f;
+        double cosYaw = Math.cos(halfYaw);
+        double sinYaw = Math.sin(halfYaw);
+        double cosPitch = Math.cos(halfPitch);
+        double sinPitch = Math.sin(halfPitch);
+        double cosRoll = Math.cos(halfRoll);
+        double sinRoll = Math.sin(halfRoll);
+        x = cosRoll * sinPitch * cosYaw + sinRoll * cosPitch * sinYaw;
+        y = cosRoll * cosPitch * sinYaw - sinRoll * sinPitch * cosYaw;
+        z = sinRoll * cosPitch * cosYaw - cosRoll * sinPitch * sinYaw;
+        w = cosRoll * cosPitch * cosYaw + sinRoll * sinPitch * sinYaw;
+    }
+
+    @Override
+    public double getX() {
+        return x;
+    }
+
+    @Override
+    public double getY() {
+        return y;
+    }
+
+    @Override
+    public double getZ() {
+        return z;
+    }
+
+    @Override
+    public double getW() {
+        return w;
+    }
+
+    /**
+     * Sets the value of this quaternion to the conjugate of quaternion q1.
+     * @param q1 the source vector
+     */
+    public final void conjugate(Quat4d q1) {
+        this.x = -q1.x;
+        this.y = -q1.y;
+        this.z = -q1.z;
+        this.w = q1.w;
+    }
+
+    /**
+     * Negate the value of of each of this quaternion's x,y,z coordinates in place.
+     */
+    public final void conjugate() {
+        this.x = -this.x;
+        this.y = -this.y;
+        this.z = -this.z;
+    }
+
+    /**
+     * @param v the vector to multiply with
+     */
+    public void mul(Vector3f v) {
+        double rx = w * v.getX() + y * v.getZ() - z * v.getY();
+        double ry = w * v.getY() + z * v.getX() - x * v.getZ();
+        double rz = w * v.getZ() + x * v.getY() - y * v.getX();
+        this.w = -x * v.getX() - y * v.getY() - z * v.getZ();
+        this.x = rx;
+        this.y = ry;
+        this.z = rz;
+    }
+
+    /**
+      * Sets the value of this quaternion to the quaternion product of
+      * itself and q1 (this = this * q1).  
+      * @param q1 the other quaternion
+      */
+    public final void mul(Quat4d q1) {
+        double nw = this.w * q1.w - this.x * q1.x - this.y * q1.y - this.z * q1.z;
+        double nx = this.w * q1.x + q1.w * this.x + this.y * q1.z - this.z * q1.y;
+        double ny = this.w * q1.y + q1.w * this.y - this.x * q1.z + this.z * q1.x;
+        this.z = this.w * q1.z + q1.w * this.z + this.x * q1.y - this.y * q1.x;
+        this.w = nw;
+        this.x = nx;
+        this.y = ny;
+    }
+
+    /**
+      * Multiplies this quaternion by the inverse of quaternion q1 and places
+      * the value into this quaternion.  The value of the argument quaternion
+      * is preserved (this = this * q^-1).
+      * @param q1 the other quaternion
+      */
+    public final void mulInverse(Quat4d q1) {
+        Quat4d tempQuat = new Quat4d(q1);
+
+        tempQuat.inverse();
+        this.mul(tempQuat);
+    }
+
+    /**
+     * Invert this quaternion
+     */
+    public final void inverse() {
+        x = -x;
+        y = -y;
+        z = -z;
+    }
+
+    /**
+     * Normalizes the value of this quaternion in place.
+     */
+    public final void normalize() {
+        double norm = (this.x * this.x + this.y * this.y + this.z * this.z + this.w * this.w);
+
+        if (norm > 0.0) {
+            norm = 1.0 / Math.sqrt(norm);
+            this.x *= norm;
+            this.y *= norm;
+            this.z *= norm;
+            this.w *= norm;
         } else {
-            q1x = q1.getX();
-            q1Y = q1.getY();
-            q1Z = q1.getZ();
-            q1W = q1.getW();
+            this.x = 0.0;
+            this.y = 0.0;
+            this.z = 0.0;
+            this.w = 0.0;
         }
-
-        double s1;
-        double s2;
-
-        if ((1.0 - dot) > EPS) {
-            double om = Math.acos(dot);
-            double sinom = Math.sin(om);
-            s1 = Math.sin((1.0 - alpha) * om) / sinom;
-            s2 = Math.sin(alpha * om) / sinom;
-        } else {
-            s1 = 1.0 - alpha;
-            s2 = alpha;
-        }
-        
-        double w = s1 * q1W + s2 * q2.getW();
-        double x = s1 * q1x + s2 * q2.getX();
-        double y = s1 * q1Y + s2 * q2.getY();
-        double z = s1 * q1Z + s2 * q2.getZ();
-        
-        return new MutableQuat4d(x, y, z, w);
-    }
-    
-    /**
-     * All implementations with the same coordinates are equal
-     */
-    @Override
-    public final boolean equals(Object obj) {
-        if (this == obj) {
-            return true;
-        }
-        if (obj instanceof Quat4d) {
-            Quat4d other = (Quat4d) obj;
-            return Double.doubleToLongBits(getX()) == Double.doubleToLongBits(other.getX())
-                && Double.doubleToLongBits(getY()) == Double.doubleToLongBits(other.getY())
-                && Double.doubleToLongBits(getZ()) == Double.doubleToLongBits(other.getZ())
-                && Double.doubleToLongBits(getW()) == Double.doubleToLongBits(other.getW());
-        }
-        return false;
     }
 
-    /**
-     * All implementations with the same coordinates have the same hash code
-     */
-    @Override
-    public final int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        long temp;
-        temp = Double.doubleToLongBits(getX());
-        result = prime * result + (int) (temp ^ (temp >>> 32));
-        temp = Double.doubleToLongBits(getY());
-        result = prime * result + (int) (temp ^ (temp >>> 32));
-        temp = Double.doubleToLongBits(getZ());
-        result = prime * result + (int) (temp ^ (temp >>> 32));
-        temp = Double.doubleToLongBits(getW());
-        result = prime * result + (int) (temp ^ (temp >>> 32));
-
-        return result;
-    }
-
-    @Override
-    public String toString() {
-        return "(" + getX() + ", " + getY() + ", " + getZ() + ", " + getW() + ")";
-    }
 }
-
